Rasa_Chatbot_Counseling/Rasa_Chatbot_Counseling/requirements.txt
absl-py==1.3.0
aio-pika==8.2.3
aiofiles==23.2.1
aiogram==2.25.2
aiohttp==3.8.6
aiormq==6.4.2
aiosignal==1.3.1
APScheduler==3.9.1.post1
astunparse==1.6.3
async-timeout==4.0.3
asynctest==0.13.0
attrs==22.1.0
Babel==2.9.1
backports.zoneinfo==0.2.1
bidict==0.22.1
blis==0.7.11
boto3==1.33.13
botocore==1.33.13
CacheControl==0.12.14
cachetools==5.3.2
catalogue==2.0.10
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==2.2.1
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confection==0.1.4
confluent-kafka==1.9.2
cryptography==42.0.2
cycler==0.11.0
dask==2022.2.0
dnspython==1.16.0
docopt==0.6.2
fbmessenger==6.0.0
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
gast==0.4.0
google-auth==2.27.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
h5py==3.8.0
httptools==0.6.0
humanfriendly==10.0
idna==3.6
importlib-metadata==6.7.0
importlib-resources==5.12.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.2.0
jsonpickle==3.0.2
jsonschema==4.17.3
keras==2.11.0
kiwisolver==1.4.5
ko-core-news-md==3.7.0
langcodes==3.3.0
libclang==16.0.6
locket==1.0.0
magic-filter==1.0.12
Markdown==3.4.4
MarkupSafe==2.1.4
matplotlib==3.5.3
mattermostwrapper==2.2
msgpack==1.0.5
multidict==5.2.0
murmurhash==1.0.10
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==20.9
pamqp==3.2.1
partd==1.4.1
Pillow==9.5.0
pip==22.3.1
pkgutil_resolve_name==1.3.10
pluggy==1.2.0
portalocker==2.7.0
preshed==3.0.9
prompt-toolkit==3.0.28
protobuf==3.19.6
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.2
pydot==1.4.2
PyJWT==2.8.0
pykwalify==1.8.0
pymongo==3.10.1
pyparsing==3.1.1
pyreadline==2.1
pyrsistent==0.19.3
python-crfsuite==0.9.10
python-dateutil==2.8.2
python-engineio==4.8.2
python-socketio==5.11.0
pytz==2022.7.1
pywin32==306
PyYAML==6.0.1
questionary==1.10.0
randomname==0.1.5
rasa==3.5.17
rasa-sdk==3.5.1
redis==4.6.0
regex==2022.10.31
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rocketchat-API==1.28.1
rsa==4.9
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
s3transfer==0.8.2
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.0.2
scipy==1.7.3
sentry-sdk==1.14.0
setuptools==65.6.3
simple-websocket==1.0.0
six==1.16.0
sklearn-crfsuite==0.3.6
slack_sdk==3.26.2
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.51
srsly==2.4.8
tabulate==0.9.0
tarsafe==0.0.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-addons==0.19.0
tensorflow-estimator==2.11.0
tensorflow-hub==0.12.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3
